Update register remote service API
Change `RegisterRemoteSKI` api call with a 2nd argument to provide the SHIP-ID if known. If provided (as a non empty string), upon SHIP handshake it will check the reported SHIP-ID from the remote service, and if it differs, will close the connection. The SHIP-ID is also made available via mDNS, and therefor should be known at all times.

The implementation of `ServiceShipIDUpdate` of `HubReaderInterface` is called to report the SHIP-ID, if it wasn’t known yet.